echoAI/requirements.txt

# Core framework
fastapi>=0.110
uvicorn[standard]>=0.23
pydantic>=2.5
pyjwt>=2.8
httpx>=0.27
python-multipart>=0.0.9
python-dotenv>=1.0.0

# LLM and workflow orchestration
openai>=1.0.0
anthropic>=0.18.0
langgraph>=0.0.20
langchain-core>=0.1.0
langchain-openai>=0.1.0

# JSON schema validation
jsonschema>=4.17.0
